requirements.txt
# LLM + Agent framework
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.0.10
openai>=1.12.0

# Data validation & processing
pydantic>=2.6
pandas>=2.2
typing-extensions>=4.9

# PDF parsing
PyMuPDF>=1.23
pdfplumber>=0.10

# Environment & utilities
python-dotenv>=1.0
tqdm>=4.66